refactor: return id when creating
previously, creating 2 tasks in the same session didn't actually increment their ID, so marking would be incorrect. we now return id when creating, and we set the id so that id is up to date

+++ src/main/java/Managers/TaskManager.java
@@ -127,7 +127,7 @@ public class TaskManager {
     Task task = this.get(taskIndex);
     this.printSeparator();
     TaskType type = getTypeOfTask(task.toString());
-    task = TaskDao.mark(task.getId(), type.getCommand(), task, isDone);
+    task = TaskDao.mark((int)task.getId(), type.getCommand(), task, isDone);
     System.out.println("Nice! I've marked this task as done: \n" + task);
     this.printSeparator();
   }
